refactor: modularize automation adapters into sub-modules and add X-API adapter tests.
