--- roadmap/artifacts/session-09-conformance-results.md
@@ -1,35 +1,0 @@
-# Session 09 — Kernel Conformance Results
-
-**Generated:** 2026-02-28 00:09 UTC
-
-**Conformance rate:** 27/27 (100.0%)
-
-| Tool | Conformant |
-|------|------------|
-| get_tweet | PASS |
-| get_user_by_username | PASS |
-| search_tweets | PASS |
-| get_user_mentions | PASS |
-| get_user_tweets | PASS |
-| get_home_timeline | PASS |
-| get_me | PASS |
-| get_followers | PASS |
-| get_following | PASS |
-| get_user_by_id | PASS |
-| get_liked_tweets | PASS |
-| get_bookmarks | PASS |
-| get_users_by_ids | PASS |
-| get_tweet_liking_users | PASS |
-| post_tweet | PASS |
-| reply_to_tweet | PASS |
-| quote_tweet | PASS |
-| delete_tweet | PASS |
-| post_thread | PASS |
-| like_tweet | PASS |
-| unlike_tweet | PASS |
-| follow_user | PASS |
-| unfollow_user | PASS |
-| retweet | PASS |
-| unretweet | PASS |
-| bookmark_tweet | PASS |
-| unbookmark_tweet | PASS |
