--- roadmap/artifacts/session-09-handoff.md
@@ -1,38 +1,0 @@
-# Session 09 — Handoff
-
-**Generated:** 2026-02-28 00:09 UTC
-
-## Scenarios
-
-| Scenario | Description | Steps | Total (ms) | Success | Schema |
-|----------|-------------|-------|------------|---------|--------|
-| D | Direct kernel read flow: get_tweet, search, followers, me | 4 | 0 | PASS | PASS |
-| E | Mutation with idempotency enforcement | 3 | 4 | PASS | PASS |
-| F | Rate-limited and auth error behavior validation | 2 | 0 | PASS | PASS |
-| G | Provider switching: MockProvider vs ScraperReadProvider | 3 | 0 | PASS | PASS |
-
-## Quality Gates
-
-| Gate | Rate | Threshold | Status |
-|------|------|-----------|--------|
-| Schema validation | 100.0% | 95% | PASS |
-| Unknown errors | 0.0% | 5% | PASS |
-| Kernel conformance | 100.0% | 100% | PASS |
-| Error code accuracy | 100.0% | 100% | PASS |
-
-**Overall: ALL GATES PASS**
-
-## Session 09 Artifacts
-
-- `session-09-conformance-results.md` — kernel tool conformance
-- `session-09-golden-fixtures.json` — schema golden fixtures
-- `session-09-schema-golden-report.md` — golden fixture report
-- `session-09-eval-results.json` — eval scenario results
-- `session-09-latency-report.md` — benchmark latency gates
-
-## What Session 10 Must Finalize
-
-1. Release documentation (README, CHANGELOG, API docs)
-2. Final manifest regeneration (`cargo test -p tuitbot-mcp manifest -- --ignored`)
-3. Version bump and crates.io publish preparation
-4. End-to-end integration test with real X API sandbox (if available)
