--- roadmap/artifacts/session-09-schema-golden-report.md
@@ -1,16 +1,0 @@
-# Session 09 — Schema Golden Report
-
-**Generated:** 2026-02-28 00:09 UTC
-
-| Family | Tools | Keys | Pagination | Status |
-|--------|-------|------|------------|--------|
-| engage_result | 3 | 2 | no | PASS |
-| error_auth | 1 | 3 | no | PASS |
-| error_rate_limited | 1 | 4 | no | PASS |
-| read_single_tweet | 1 | 6 | no | PASS |
-| read_single_user | 3 | 4 | no | PASS |
-| read_tweet_list | 6 | 3 | yes | PASS |
-| read_users_list | 4 | 2 | yes | PASS |
-| write_result | 3 | 2 | no | PASS |
-
-**Total families:** 8
