--- roadmap/artifacts/task-01-baseline-benchmark.json
@@ -1,60 +1,0 @@
-{
-  "aggregate": {
-    "max_ms": "2.881",
-    "min_ms": "0.054",
-    "p50_ms": "0.559",
-    "p95_ms": "1.432"
-  },
-  "benchmark": "task-01-baseline",
-  "iterations_per_tool": 5,
-  "migrated_tools": 5,
-  "schema_pass_rate": "100%",
-  "tools": [
-    {
-      "avg_ms": "0.937",
-      "iterations": 5,
-      "max_ms": "1.432",
-      "min_ms": "0.691",
-      "p50_ms": "0.894",
-      "p95_ms": "1.432",
-      "tool": "get_capabilities"
-    },
-    {
-      "avg_ms": "0.277",
-      "iterations": 5,
-      "max_ms": "0.559",
-      "min_ms": "0.176",
-      "p50_ms": "0.208",
-      "p95_ms": "0.559",
-      "tool": "health_check"
-    },
-    {
-      "avg_ms": "1.524",
-      "iterations": 5,
-      "max_ms": "2.881",
-      "min_ms": "1.022",
-      "p50_ms": "1.280",
-      "p95_ms": "2.881",
-      "tool": "get_stats"
-    },
-    {
-      "avg_ms": "0.337",
-      "iterations": 5,
-      "max_ms": "1.154",
-      "min_ms": "0.109",
-      "p50_ms": "0.142",
-      "p95_ms": "1.154",
-      "tool": "list_pending"
-    },
-    {
-      "avg_ms": "0.235",
-      "iterations": 5,
-      "max_ms": "0.897",
-      "min_ms": "0.054",
-      "p50_ms": "0.069",
-      "p95_ms": "0.897",
-      "tool": "list_unreplied_tweets_with_limit"
-    }
-  ],
-  "total_tools": 27
-}
